vfs_linux_arm64: adjust for libc@v1.52.1, updates #185

+++ vfs/vfs_linux_arm64.go
@@ -1355,7 +1355,7 @@ func vfsWrite(tls *libc.TLS, pFile uintptr, zBuf uintptr, iAmt int32, iOfst sqli
 			}
 			if (*VFSFile)(unsafe.Pointer(p)).nBuffer == 0 || (*VFSFile)(unsafe.Pointer(p)).iBufferOfst+sqlite3_int64((*VFSFile)(unsafe.Pointer(p)).nBuffer) == i {
 			} else {
-				libc.X__assert_fail(tls, ts+43, ts+89, uint32(294), uintptr(unsafe.Pointer(&__func__4)))
+				libc.X__assert_fail(tls, ts+43, ts+89, 294, uintptr(unsafe.Pointer(&__func__4)))
 			}
 			(*VFSFile)(unsafe.Pointer(p)).iBufferOfst = i - sqlite3_int64((*VFSFile)(unsafe.Pointer(p)).nBuffer)
